Reorganize runtime support into a single module
