+++ tests/namespace_test.py
@@ -1,25 +1,31 @@
+from unittest.mock import patch
+
 import pyrsistent
 import pytest
+
 import basilisp.lang.atom as atom
-import basilisp.lang.namespace as namespace
+import basilisp.lang.runtime as runtime
 import basilisp.lang.symbol as sym
+from basilisp.lang.runtime import Namespace
 
 
 @pytest.fixture
 def core_ns_sym():
-    return sym.symbol(namespace._CORE_NS)
+    return sym.symbol(runtime._CORE_NS)
 
 
 @pytest.fixture
 def core_ns(core_ns_sym):
-    ns = namespace.Namespace(core_ns_sym)
+    ns = Namespace(core_ns_sym)
     return ns
 
 
 @pytest.fixture
 def ns_cache(core_ns_sym: sym.Symbol,
-             core_ns: namespace.Namespace) -> atom.Atom:
-    return atom.Atom(pyrsistent.pmap({core_ns_sym: core_ns}))
+             core_ns: Namespace) -> patch:
+    """Patch the Namespace cache with a test fixture."""
+    return patch('basilisp.lang.runtime.Namespace._NAMESPACES',
+                 new=atom.Atom(pyrsistent.pmap({core_ns_sym: core_ns})))
 
 
 @pytest.fixture
@@ -27,39 +33,44 @@ def ns_sym() -> sym.Symbol:
     return sym.symbol("some.ns")
 
 
-def test_create_ns(ns_sym: sym.Symbol, ns_cache: atom.Atom):
-    assert len(ns_cache.deref().keys()) == 1
-    ns = namespace.get_or_create(ns_sym, ns_cache=ns_cache)
-    assert isinstance(ns, namespace.Namespace)
-    assert ns.name == ns_sym.name
-    assert len(ns_cache.deref().keys()) == 2
+def test_create_ns(ns_sym: sym.Symbol, ns_cache: patch):
+    with ns_cache as cache:
+        assert len(cache.deref().keys()) == 1
+        ns = Namespace.get_or_create(ns_sym)
+        assert isinstance(ns, Namespace)
+        assert ns.name == ns_sym.name
+        assert len(cache.deref().keys()) == 2
 
 
 @pytest.fixture
 def ns_cache_with_existing_ns(ns_sym: sym.Symbol, core_ns_sym: sym.Symbol,
-                              core_ns: namespace.Namespace) -> atom.Atom:
-    return atom.Atom(
+                              core_ns: Namespace) -> patch:
+    """Patch the Namespace cache with a test fixture with an existing namespace."""
+    return patch('basilisp.lang.runtime.Namespace._NAMESPACES',
+                 atom.Atom(
         pyrsistent.pmap({
             core_ns_sym: core_ns,
-            ns_sym: namespace.Namespace(ns_sym)
-        }))
+            ns_sym: Namespace(ns_sym)
+        })))
 
 
 def test_get_existing_ns(ns_sym: sym.Symbol,
-                         ns_cache_with_existing_ns: atom.Atom):
-    assert len(ns_cache_with_existing_ns.deref().keys()) == 2
-    ns = namespace.get_or_create(ns_sym, ns_cache=ns_cache_with_existing_ns)
-    assert isinstance(ns, namespace.Namespace)
-    assert ns.name == ns_sym.name
-    assert len(ns_cache_with_existing_ns.deref().keys()) == 2
+                         ns_cache_with_existing_ns: patch):
+    with ns_cache_with_existing_ns as cache:
+        assert len(cache.deref().keys()) == 2
+        ns = Namespace.get_or_create(ns_sym)
+        assert isinstance(ns, Namespace)
+        assert ns.name == ns_sym.name
+        assert len(cache.deref().keys()) == 2
 
 
-def test_remove_ns(ns_sym: sym.Symbol, ns_cache_with_existing_ns: atom.Atom):
-    assert len(ns_cache_with_existing_ns.deref().keys()) == 2
-    ns = namespace.remove(ns_sym, ns_cache=ns_cache_with_existing_ns)
-    assert isinstance(ns, namespace.Namespace)
-    assert ns.name == ns_sym.name
-    assert len(ns_cache_with_existing_ns.deref().keys()) == 1
+def test_remove_ns(ns_sym: sym.Symbol, ns_cache_with_existing_ns: patch):
+    with ns_cache_with_existing_ns as cache:
+        assert len(cache.deref().keys()) == 2
+        ns = Namespace.remove(ns_sym)
+        assert isinstance(ns, Namespace)
+        assert ns.name == ns_sym.name
+        assert len(cache.deref().keys()) == 1
 
 
 @pytest.fixture
@@ -68,8 +79,9 @@ def other_ns_sym() -> sym.Symbol:
 
 
 def test_remove_non_existent_ns(other_ns_sym: sym.Symbol,
-                                ns_cache_with_existing_ns: atom.Atom):
-    assert len(ns_cache_with_existing_ns.deref().keys()) == 2
-    ns = namespace.remove(other_ns_sym, ns_cache=ns_cache_with_existing_ns)
-    assert ns is None
-    assert len(ns_cache_with_existing_ns.deref().keys()) == 2
+                                ns_cache_with_existing_ns: patch):
+    with ns_cache_with_existing_ns as cache:
+        assert len(cache.deref().keys()) == 2
+        ns = Namespace.remove(other_ns_sym)
+        assert ns is None
+        assert len(cache.deref().keys()) == 2
